moving common code to model.py
